requirements.txt
asgiref==3.9.1
asttokens==3.0.0
bcrypt==4.3.0
colorama==0.4.6
crispy-bootstrap5==2025.6
decorator==5.2.1
Django==5.2.5
django-crispy-forms==2.4
django-jazzmin==3.0.1
djangorestframework==3.16.1
executing==2.2.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
mysqlclient==2.2.7
parso==0.8.5
pillow==11.3.0
prompt_toolkit==3.0.52
pure_eval==0.2.3
Pygments==2.19.2
stack-data==0.6.3
traitlets==5.14.3
tzdata==2025.2
wcwidth==0.2.13
